feat: Navigation code clean up & added navigation css module
