+++ packages/website/components/navbar/index.js
@@ -1,0 +1,2 @@
+import Navbar from './navbar'
+export default Navbar
